refactor: move split publish function
