+++ cloud/functions/src/index.ts
@@ -2,8 +2,6 @@ import * as functions from 'firebase-functions';
 
 import 'firebase-functions/lib/logger/compat';
 
-import * as cors from 'cors';
-
 import * as admin from 'firebase-admin';
 const app: admin.app.App = admin.initializeApp();
 app.firestore().settings({timestampsInSnapshots: true});
@@ -11,14 +9,12 @@ app.firestore().settings({timestampsInSnapshots: true});
 import {applyWatchDeckCreate, applyWatchDeckDelete, applyWatchDeckUpdate} from './watch/watch-deck';
 import {applyWatchUserCreate, applyWatchUserDelete, applyWatchUserUpdate} from './watch/watch-user';
 
-import {publishDeck} from './request/publish/publish-deck';
+import {publishJob} from './request/publish';
 
 const runtimeOpts = {
   timeoutSeconds: 120,
   memory: <const>'1GB',
 };
-
-const corsHandler = cors({origin: true});
 
 export const watchDeckUpdate = functions.runWith(runtimeOpts).firestore.document('decks/{deckId}').onUpdate(applyWatchDeckUpdate);
 
@@ -32,8 +28,4 @@ export const watchUserDelete = functions.auth.user().onDelete(applyWatchUserDele
 
 export const watchUserCreate = functions.auth.user().onCreate(applyWatchUserCreate);
 
-export const publish = functions.runWith(runtimeOpts).https.onRequest((request: functions.Request, response: functions.Response<any>) => {
-  corsHandler(request, response, async () => {
-    await publishDeck(request, response);
-  });
-});
+export const publish = functions.runWith(runtimeOpts).https.onRequest(publishJob);
